requirements.txt
requests
beautifulsoup4
pydub
google-generativeai
python-dotenv 